refactor : planId int -> Long 으로 변경

+++ presentation/src/main/java/com/yapp/growth/presentation/ui/main/manage/confirm/FixPlanContract.kt
@@ -31,7 +31,7 @@ class FixPlanContract {
     sealed class FixPlanSideEffect : ViewSideEffect {
         object ShowBottomSheet : FixPlanSideEffect()
         object HideBottomSheet : FixPlanSideEffect()
-        data class NavigateToNextScreen(val planId: Int) : FixPlanSideEffect()
+        data class NavigateToNextScreen(val planId: Long) : FixPlanSideEffect()
         object NavigateToPreviousScreen : FixPlanSideEffect()
     }
